* 'engine-api' of https://github.com/arangodb/arangodb:
  fix objectId handling
  Fixing windows build
  Fixing windows build
  Bugfix: we want the common _and_ the server tests.
  Conversion warnings
  Conversion errors
  Renaming test files
  fix windows build
  fix if clause
  exclude longrunning tests
  add slash here too.
  ...

+++ arangod/Aql/IndexBlock.h
@@ -162,7 +162,7 @@ class IndexBlock final : public ExecutionBlock {
 
   /// @brief Counter how many documents have been returned/skipped
   ///        during one call.
-  size_t _returned;
+  uint64_t _returned;
 
   /// @brief Collect several AQLItemsBlocks
   BlockCollector _collector;
